Support for transcoding to compliant AAC.
This does not re-encode the audio stream, and does not have the code required
to reconstruct the uncompressed audio stream. It only parses the HDC packets
and writes out the same data as normal AAC packets.

The code is based on FAAD2 so the GPL applies. It is unknown if patents apply,
but I believe not since it cannot reconstruct the audio.